#scratch_commit 1.Created first code files, and migrated some code snippets from an old scratch file with some refactoring. 2.Developing a design of file tree structure. 3.Bit configuration of '.gitignore' for PyCharm Editor.
Note: Declaring 3 types of commits - 'major', 'minor', and 'scratch'. First two types of commits are considered as legit ones, probably with some obvious changes or improvements, whereas the other one may be not. The 'scratch commits' are temporary, or are used for unclear situations and times - so (note) that they would be pruned while clearing legacies in VCS.